Features + bug fixes
1) Fixed runways coordinates
2) Changed Menu background
3) Added Creator text to render in window
4) Changed Math.h return types from float to double
5) Menu has been reworked
6) Fixed Path::GetPointPosition() to return (0,0) if point is null

+++ 2D_ATC/Math.h
@@ -5,16 +5,16 @@
 
 #define PI 3.14159265
 
-float DistanceToPoint(sf::Vector2f pointA, sf::Vector2f pointB)
+double DistanceToPoint(sf::Vector2f pointA, sf::Vector2f pointB)
 {
-	float distance = sqrt(pow(pointB.x-pointA.x,2)+pow(pointB.y-pointA.y,2));
+	double distance = sqrt(pow(pointB.x-pointA.x,2)+pow(pointB.y-pointA.y,2));
 
 	return distance;
 }
 
-float DirectionToPoint(sf::Vector2f pointA, sf::Vector2f pointB)
+double DirectionToPoint(sf::Vector2f pointA, sf::Vector2f pointB)
 {
-	float d = atan2(pointA.y - pointB.y, pointA.x - pointB.x);
+	double d = atan2(pointA.y - pointB.y, pointA.x - pointB.x);
 	d *= 180 / PI;
 	d -= 90;
 	if (d < 0)
